Adding migration, seeder and nodel updated
