+++ .env.example
@@ -8,10 +8,10 @@ DEV_PASSWORD=postgres
 DEV_HOST=localhost
 DEV_PORT=5432
 # Redis Configuration
-REDIS_HOST=your_redis_host
-REDIS_PORT=your_redis_port
-REDIS_PASSWORD=your_redis_password
-REDIS_DB=your_redis_database
+REDIS_HOST=redis-13935.c83.us-east-1-2.ec2.redns.redis-cloud.com
+REDIS_PORT=13935
+REDIS_PASSWORD=kR6SRYY2PlmHSRMyKT8d4HH7O
+REDIS_DB=0
 # JWT Configuration
 JWT_SECRET=your-super-secret-jwt-key-change-in-production
 CLOUDINARY_API_KEY=your-cloudinary-api-key
