+++ package.json
@@ -42,6 +42,7 @@
   "devDependencies": {
     "@eslint/js": "^9.32.0",
     "@types/express": "^5.0.3",
+    "@types/ioredis": "^4.28.10",
     "@types/morgan": "^1.9.10",
     "@types/multer": "^2.0.0",
     "@types/node": "^24.2.0",
